requirements.txt
bokeh>=1.2
colorcet
datajoint>=0.12
GitPython
python-dotenv
holoviews
jupyter-dash
matplotlib
neo>=0.8.0
numpy>=1.16.4
pandas>=0.24.2
plotly>=4.1.0
scikit-learn>=0.21.2
scipy>=1.3.0
seaborn>=0.10.0
swatch